+++ R/new_package.R
@@ -814,7 +814,9 @@ new_package <- function(
 
   ui_title("Adding Badges to README")
 
-  add_cran_badge(quiet = quiet)
+  add_cran_version_badge(quiet = quiet)
+  add_cran_total_download_badge(quiet = quiet)
+  add_cran_downloads_badge(quiet = quiet)
 
   if (gh_check) {
     add_github_actions_check_badge(organisation, quiet = quiet)